Home Page
- Changed some names
- Added Search bar
- Restructured the .yaml file
- Created three states in home state
- Created some methods in the WeatherForeCastModel
- Adjusted the models to work with Double and not Int
- Built the main structure for the HomePage

+++ lib/models/weather_forecast_model/current.dart
@@ -14,7 +14,7 @@ class Current with _$Current {
   factory Current({
     @JsonKey(name: 'last_updated_epoch') int? lastUpdatedEpoch,
     @JsonKey(name: 'last_updated') String? lastUpdated,
-    @JsonKey(name: 'temp_c') int? tempC,
+    @JsonKey(name: 'temp_c') double? tempC,
     @JsonKey(name: 'temp_f') double? tempF,
     @JsonKey(name: 'is_day') int? isDay,
     Condition? condition,
